Mid-fix on Anonymizer refactor
The Anonymizer's biggest weakness is that it does not systematically
work through unique identifiers presented - instead it works with
unique-identifiers as they are read from the file.

This code refactor creates a complete list of unique  identifiers and
then works through them based on a size ordering.

Stil TODO:
Check that unique found names that are smaller would also refactor to
the larger name given current configuration