Refactor to Anonymizer complete
Forced ordering of checks, and also added a symmetry condition - a name
can only be replaced by another name if that name would also think the
new name is a good replacement.  So, assuming that three letter words
are an exact match, Jane and Jan would never replace, because while Jane
might think Jan is a good substitution, Jan would not.