Added Review Panels for Cleaned Names and Cleaned Roles
Tool now creates explicit tables of cleaning names and roles for user
review outside of the main log.  It creates these on new tabbed panels
"Review Names" and "Review Roles".  These panels are intended to help
the user be certain the cleaning process is doing what they expect and
want.